I just have structured folders and files this way which benefits project management, collaboration, and maintainability.
