+++ app/web_app/templates/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Phishing Detection Tool</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
+</head>
+<body>
+    <div class="container">
+        <h2>Phishing Detection Tool</h2>
+        <form action="/submit" method="post">
+            <label for="url">URL:</label><br>
+            <input type="text" id="url" name="url"><br><br>
+            <label for="text">Email/Text Content:</label><br>
+            <textarea id="text" name="text" rows="10" cols="60"></textarea><br><br>
+            <input type="submit" value="Analyze">
+        </form>
+    </div>
+</body>
+</html>
