--- templates/result.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Analysis Result</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
-</head>
-<body>
-    <div class="result-box">
-        {% if result.is_suspicious %}
-            <h2 class="warning">⚠️ This message is {{ result.score }}% suspicious</h2>
-            <img src="{{ url_for('static', filename='warning.png') }}" width="100" alt="Warning">
-            <p class="be-careful">Be careful! This might be a phishing attempt.</p>
-            <ul>
-                {% for item in result.findings %}
-                    <li>{{ item }}</li>
-                {% endfor %}
-            </ul>
-        {% else %}
-            <h2 class="safe">✅ This message appears to be safe</h2>
-        {% endif %}
-        <a href="/">← Back to Form</a>
-    </div>
-</body>
-</html>
